+++ src/api/compile.rs
@@ -2,7 +2,6 @@
 
 use crate::func::native::locked_write;
 use crate::parser::{ParseResult, ParseState};
-use crate::tokenizer::lex_raw;
 use crate::types::StringsInterner;
 use crate::{Engine, OptimizationLevel, Scope, AST};
 #[cfg(feature = "no_std")]
@@ -219,7 +218,7 @@ impl Engine {
         scripts: impl AsRef<[S]>,
         optimization_level: OptimizationLevel,
     ) -> ParseResult<AST> {
-        let (stream, tc) = lex_raw(self, scripts.as_ref(), self.token_mapper.as_deref());
+        let (stream, tc) = self.lex(scripts.as_ref());
 
         let mut interner;
         let mut guard;
@@ -303,7 +302,7 @@ impl Engine {
         script: impl AsRef<str>,
     ) -> ParseResult<AST> {
         let scripts = [script];
-        let (stream, t) = lex_raw(self, &scripts, self.token_mapper.as_deref());
+        let (stream, t) = self.lex(&scripts);
 
         let mut interner;
         let mut guard;