+++ src/func/call.rs
@@ -7,7 +7,7 @@ use crate::engine::{
     KEYWORD_DEBUG, KEYWORD_EVAL, KEYWORD_FN_PTR, KEYWORD_FN_PTR_CALL, KEYWORD_FN_PTR_CURRY,
     KEYWORD_IS_DEF_VAR, KEYWORD_PRINT, KEYWORD_TYPE_OF,
 };
-use crate::eval::{search_namespace, Caches, FnResolutionCacheEntry, GlobalRuntimeState};
+use crate::eval::{Caches, FnResolutionCacheEntry, GlobalRuntimeState};
 use crate::tokenizer::{is_valid_function_name, Token};
 use crate::types::dynamic::Union;
 use crate::{
@@ -1355,7 +1355,7 @@ impl Engine {
                 }
 
                 let mut target =
-                    search_namespace(self, global, caches, scope, this_ptr, first_expr)?;
+                    self.search_namespace(global, caches, scope, this_ptr, first_expr)?;
 
                 if target.as_ref().is_read_only() {
                     target = target.into_owned();
@@ -1452,7 +1452,7 @@ impl Engine {
                     arg_values.push(value.flatten());
                 }
 
-                let target = search_namespace(self, global, caches, scope, this_ptr, first_expr)?;
+                let target = self.search_namespace(global, caches, scope, this_ptr, first_expr)?;
 
                 if target.is_shared() || target.is_temp_value() {
                     arg_values[0] = target.take_or_clone().flatten();
@@ -1480,7 +1480,8 @@ impl Engine {
         }
 
         // Search for the root namespace
-        let module = crate::eval::search_imports(self, global, namespace)
+        let module = self
+            .search_imports(global, namespace)
             .ok_or_else(|| ERR::ErrorModuleNotFound(namespace.to_string(), namespace.position()))?;
 
         // First search script-defined functions in namespace (can override built-in)