+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Enhancements
 
 * `FuncRegistration::in_global_namespace` and `FuncRegistration::in_internal_namespace` are added to avoid pulling in `FnNamespace`.
 * Array/BLOB/string iterators are defined also within the `BasicIteratorPackage` in addition to the regular array/BLOB/string packages.
+* `LexError::Runtime` is added for use with `Engine::on_parse_token`.
 
 
 Version 1.17.2